Refactored out the BindingData into a separate header file (according to the "one class per file" paradigm), and moved the setting up of the Javascript methods into a separate helper method. This frees us from one (quite superfluous) comment, which is always a good thing.
